+++ app/helpers/same_articles_helper.rb
@@ -1,5 +1,6 @@
 module SameArticlesHelper
   def rand_article(articles)
-    articles.where("category_id = ? AND id != ?", params[:category_id], params[:id]).order("RANDOM()").limit(3)
+    current_article = articles.find(params[:id])
+    articles.where("category_id = ? OR subject = ? AND id != ?", params[:category_id], current_article.subject, current_article.id).order("RANDOM()").limit(3)
   end
 end
