+++ app/views/articles/edit.html.slim
@@ -1,5 +1,1 @@
-.container.z-ind
-  .row.justify-content-center.my-5
-    .col-6
-      h1 Редагувати статтю
-      = render partial: "articles/form", locals: { path: category_article_path  }
+= render partial: "articles/form", locals: { path: category_article_path }
