+++ app/views/articles/new.html.slim
@@ -1,6 +1,1 @@
-.container.z-ind
-  .row.justify-content-center.my-5
-    .col-6
-      h1 Створити нову статтю
-
-      = render partial: "articles/form", locals: { path: @article }
+= render partial: "articles/form", locals: { path: @article }
