+++ app/views/articles/show.html.slim
@@ -24,7 +24,7 @@
       .card.rounded-6.bg-0
         =image_tag @article.image.url, class: "card-img-top" if @article.image.url
         .card-body
-          p.text-descr.pb-2 = @article.description.html_safe
+          .text-descr.pb-2.select-text = @article.description.html_safe
           p.share-header.pb-1 Теги
           span.subject.article-tag.me-2 = @article.subject
           span.category.article-tag = @article.category.name
@@ -38,7 +38,7 @@
                 = link_to edit_category_article_path, class: "btn main-color" do
                   i.fas.fa-pen
 
-                = link_to category_article_path(@article.category, @article), class: "btn main-color", method: :delete do
+                = link_to category_article_path(@article.category, @article), class: "btn main-color", remote: true, method: :delete do
                   i.fas.fa-trash
 
     .col-lg-3.col-12
